requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy==2.0.23
aiosqlite==0.19.0
aiohttp==3.9.1
python-dotenv==1.0.0
pydantic==2.5.0
python-multipart==0.0.6